+++ src/koality/executor.py
@@ -219,6 +219,12 @@ class CheckExecutor:
                 for check_config in check_bundle.checks:
                     check_factory = CHECK_MAP[check_config.check_type]
                     check_kwargs = check_config.model_dump(exclude={"check_type"}, exclude_none=True)
+
+                    # IqrOutlierCheck does not accept threshold arguments, so we remove them if they exist
+                    if check_config.check_type == "IqrOutlierCheck":
+                        check_kwargs.pop("lower_threshold", None)
+                        check_kwargs.pop("upper_threshold", None)
+
                     check_kwargs["database_accessor"] = self.config.database_accessor
                     check_kwargs["database_provider"] = self.database_provider
                     check_kwargs["identifier_format"] = self.config.defaults.identifier_format